pyproject.toml
[project]
name = "plateshapez"
version = "0.1.0"
description = "Dataset generator for adversarial plate overlay"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pyyaml>=6.0",
    "rich>=14.1.0",
    "typer>=0.16.1",
]

[tool.uv]
package = true

[project.scripts]
# Run with: uv run dev <command>
dev = "plateshapez.dev:main"
# CLI entry point as specified in project spec
advplate = "plateshapez.__main__:main"

[dependency-groups]
# Install with: uv sync --group dev
dev = [
  "ruff>=0.6",
  "mypy>=1.10",
  "pre-commit>=3.7",
  "types-pillow>=10.2.0.20240822",
  "pytest>=8.4.1",
]

[tool.ruff]
# Match project Python version
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "I",   # isort (imports)
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
check_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
pretty = true
